WIP #561: Add log_path option to store log
This commit changes the way the Tee in Pronterface works. When no log is
specified, sys.stdout is redirected as before to a tee between the
console panel and the true sys.stdout. When a log is specified,
sys.stdout is untouched and logging output is saved in both the console
panel and the log file.
The side effect is that all output must now go through the logging
module to properly get in the console panel and in the log.
Now we have to catch all "print" statements =)